# backend/src/main/java/com/HR_Managment_System/demo/dto/AuthResponse.java
package com.HR_Managment_System.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthResponse {

    private String token;
    private String email;
    private String role;
    private String message;
}