# submission/openapi.yaml
openapi: "3.0.3"
info:
  title: HR Management System API
  version: "2.0.0"
  description: |
    REST API for the Stella Technology HR Management System.

    ## Roles
    | Role | Access |
    |------|--------|
    | **ADMIN** | Registers employees & departments, views all leaves and payroll (read-only on those), approves HR employees' leaves |
    | **HR** | Manages their department only — approves employee leaves, generates payroll; submits own leaves (approved by Admin) |
    | **EMPLOYEE** | Applies for leave, views own leaves, payroll, and profile |

    ## Authentication
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click **Authorize** → enter `Bearer <token>`
    4. All subsequent requests include the token automatically

  contact:
    name: Hassan Raheem
    email: hmarwat49@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@hrms.com
        password:
          type: string
          example: Admin@1234
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: user@company.com
        password:
          type: string
          description: "Min 8 chars, requires uppercase, lowercase, and digit"
          example: Pass@1234
        role:
          type: string
          enum: [ADMIN, HR, EMPLOYEE]
    AuthResponse:
      type: object
      properties:
        token:   { type: string, example: "eyJhbGci..." }
        email:   { type: string, example: "admin@hrms.com" }
        role:    { type: string, enum: [ADMIN, HR, EMPLOYEE] }
        message: { type: string, example: "Login successful" }
    MeResponse:
      type: object
      properties:
        email: { type: string }
        role:  { type: string, enum: [ADMIN, HR, EMPLOYEE] }

    # ── Employee ──────────────────────────────────────────────────────────
...
    DepartmentRequest:
      type: object
      required: [name, hrId]
      properties:
        name:        { type: string, example: "Engineering" }
        description: { type: string, example: "Software development team" }
        managerId:   { type: string, example: "EMP-0002", description: "Optional. Must be a MANAGER-type employee." }
        hrId:        { type: string, example: "EMP-0001", description: "Required. Must be an HR-type employee not already assigned to another department." }
    DepartmentUpdateRequest:
      type: object
      description: All fields optional — only include what you want to change.
      properties:
        name:        { type: string }
        description: { type: string }
        managerId:   { type: string }
        hrId:        { type: string }
    DepartmentResponse:
      type: object
      properties:
        departmentId: { type: string,  example: "DEPT-0001" }
        name:         { type: string,  example: "Engineering" }
        description:  { type: string }
        managerId:    { type: string,  nullable: true }
        hrId:         { type: string }
        isSystem:     { type: boolean, description: "True only for the Unassigned system department (DEPT-0000)" }
        createdAt:    { type: string,  format: date-time }
        updatedAt:    { type: string,  format: date-time }

    # ── Leave ─────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:    { type: integer, example: 400 }
        message:   { type: string,  example: "Validation failed" }
        timestamp: { type: string,  format: date-time }
  responses:
    Unauthorized:
      description: "401 — Missing or expired JWT token"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: "403 — Authenticated but role does not have permission"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: "404 — Resource not found"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: "400 — Validation error or business rule violation"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - BearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: Authenticates a user and returns a JWT token. Login is case-insensitive on email. No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Login successful — returns JWT token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              example:
                token: "eyJhbGci..."
                email: "admin@hrms.com"
                role: "ADMIN"
                message: "Login successful"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register user account (Admin only)
      description: Creates a standalone user account (no employee profile). Typically not needed — registering via `POST /api/employees` creates both the employee profile and the user account in one step.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        "201":
          description: User account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Returns the email and role of the currently authenticated user. Accessible by all roles.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
              example:
                email: "admin@hrms.com"
                role: "ADMIN"
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ════════════════════════════════════════════════════════════
  #  EMPLOYEES
  # ════════════════════════════════════════════════════════════
...
  /api/departments:
    get:
      tags: [Departments]
      summary: List all departments
      description: Returns all departments. Optional name search. Accessible by ADMIN and HR.
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          schema: { type: string, example: "Engineering" }
          description: Search by department name (partial match)
      responses:
        "200":
          description: Department list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DepartmentResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Departments]
      summary: Create department
      description: |
        Creates a new department and assigns an HR to it.
        - `hrId` is required and must reference an HR-type employee.
        - The same HR cannot be assigned to more than one department.
        - `managerId` is optional and must reference a MANAGER-type employee.
        Requires **ADMIN**.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentRequest' }
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/departments/{departmentId}:
    get:
      tags: [Departments]
      summary: Get department by ID
      description: Returns a single department. Accessible by ADMIN and HR.
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema: { type: string, example: "DEPT-0001" }
      responses:
        "200":
          description: Department found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      tags: [Departments]
      summary: Update department
      description: Partially updates a department. All fields are optional. Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema: { type: string, example: "DEPT-0001" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentUpdateRequest' }
            example:
              description: "Updated team description"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Departments]
      summary: Delete department
      description: |
        Deletes a department. If employees remain, provide `targetDepartmentId` to move them first,
        or use `PUT /{departmentId}/reassign-employees` beforehand.
        Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema: { type: string, example: "DEPT-0001" }
        - name: targetDepartmentId
          in: query
          required: false
          schema: { type: string, example: "DEPT-0002" }
          description: Move employees to this department before deletion
      responses:
        "204":
          description: Department deleted (no body)
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/departments/{departmentId}/reassign-employees:
    put:
      tags: [Departments]
      summary: Reassign all employees to another department
      description: Moves all employees from this department to `targetDepartmentId`. Useful before deletion. Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema: { type: string, example: "DEPT-0001" }
        - name: targetDepartmentId
          in: query
          required: true
          schema: { type: string, example: "DEPT-0002" }
      responses:
        "204":
          description: Employees reassigned (no body)
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/departments/{departmentId}/employee-count:
    get:
      tags: [Departments]
      summary: Get employee count for a department
      description: Returns the total number of employees in the department. Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: path
          required: true
          schema: { type: string, example: "DEPT-0001" }
      responses:
        "200":
          description: Employee count
          content:
            application/json:
              schema: { type: integer, example: 12 }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════
  #  LEAVES
  # ════════════════════════════════════════════════════════════
...